% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_midlist.R
\name{extract.midlist}
\alias{extract.midlist}
\alias{[.midlist}
\alias{[.midrib}
\alias{[[.midrib}
\title{Subset MID Objects}
\usage{
\method{[}{midlist}(x, i, drop = if (missing(i)) TRUE else length(i) == 1L)

\method{[}{midrib}(x, i, drop = if (missing(i)) TRUE else length(i) == 1L)

\method{[[}{midrib}(x, i, exact = TRUE)
}
\arguments{
\item{x}{a collection object of class "midlist" or "midrib".}

\item{i}{indices specifying elements to extract. Can be numeric, character, or logical vectors.}

\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible dimension. For a collection (e.g., "mids"), extracting a single element drops it to a base object (e.g., "mid").}

\item{exact}{logical. If \code{TRUE}, exact matching is used for character strings.}
}
\value{
\code{[} returns a subsetted collection object (e.g., "midrib", "midlist") or a single base object if \code{drop = TRUE}.

\code{[[} returns a single base object (e.g., "mid").
}
\description{
S3 methods to extract parts of a "midlist" or "midrib" collection object.
}
\details{
A "midlist" or "midrib" object stores multiple objects of the same single base class: "mid", "midimp", "midcon", or "midbrk".

When extracting items using \code{[}, it returns a subsetted "midlist" or "midrib" object, preserving its collection class (e.g., "mids", "midimps").
By default, if a single base object is extracted (\code{length(i) == 1L}) and \code{drop = TRUE}, the object is simplified to a single base object (e.g., "mid", "midimp").
\code{[[} always extracts a single base object.
}
\examples{
# Fit a multivariate linear model
fit <- lm(cbind(y1, y2, y3) ~ x1 + I(x1^2), data = anscombe)

# Interpret the linear models
collection <- interpret(cbind(y1, y2, y3) ~ x1, data = anscombe, model = fit)

# Check the default labels
labels(collection)

# Rename the models in the collection
labels(collection) <- letters[1L:3L]
labels(collection)

# Extract a single base "mid" object by its new name using [[
mid <- collection[["a"]]
class(mid)

# Subset the collection to keep only the first two models using [
sub <- collection[1:2]
class(sub) # Maintains the collection class (e.g., "mids"-"midrib")
}
\seealso{
\code{\link{midlist}}, \code{\link{labels.midlist}}
}
