% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_encoder.R
\name{factor.encoder}
\alias{factor.encoder}
\alias{factor.frame}
\title{Encoder for Qualitative Variables}
\usage{
factor.encoder(
  x,
  k,
  use.catchall = TRUE,
  catchall = "(others)",
  tag = "x",
  frame = NULL,
  weights = NULL
)

factor.frame(levels, catchall = "(others)", tag = "x")
}
\arguments{
\item{x}{a vector to be encoded as a qualitative variable.}

\item{k}{an integer specifying the max number of the levels after the encoding. If not positive, all unique values of x are chosen as sample points.}

\item{use.catchall}{logical. If TRUE, less frequent levels are dropped and replaced with the catchall level.}

\item{catchall}{a character used as the name of "catchall" level for unused levels.}

\item{tag}{name of the variable.}

\item{frame}{a \code{factor.frame} object or a vector, containing the information about the levels of the variable.}

\item{weights}{optional. a numeric vector indicating the weight of each value of 'x'.}

\item{levels}{a vector to be used as the levels of the variable.}
}
\description{
Returns a list consisting of the encoding information of a predictor variable as a qualitative variable.
}
