% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_importance.R
\name{ggmid.mid.importance}
\alias{ggmid.mid.importance}
\alias{autoplot.mid.importance}
\title{Plot MID Importance with ggplot2 Package}
\usage{
\method{ggmid}{mid.importance}(
  object,
  type = c("barplot", "heatmap"),
  plot.main = TRUE,
  max.bars = NA,
  scale.palette = c("#FFFFFF", "#464646"),
  ...
)

\method{autoplot}{mid.importance}(object, ...)
}
\arguments{
\item{object}{a 'mid.importance' object to be visualized.}

\item{type}{a character, "barplot" or "heatmap", specifying the type of the plot.}

\item{plot.main}{logical. If \code{TRUE}, the main layer is not drawn.}

\item{max.bars}{an integer specifying the maximum number of bars.}

\item{scale.palette}{a character vector of length two. The color palette for the heatmap.}

\item{...}{optional parameters to be passed to the main layer.}
}
\value{
\code{ggmid.mid.importance()} returns a \code{ggplot} object.
}
\description{
For 'mid.importance' objects, \code{ggmid()} visualizes the importance of MID component functions.
}
\details{
The S3 method of \code{ggmid()} for 'mid.importance' objects creates a 'ggplot' object that visualizes the term importance of a fitted MID model.
The main layer is drawn using \code{geom_col()} or \code{geom_tile()}.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
imp <- mid.importance(mid)
ggmid(imp)
ggmid(imp, type = "heatmap")
}
