% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_midlist_conditional.R
\name{ggmid.midlist.conditional}
\alias{ggmid.midlist.conditional}
\alias{autoplot.midlist.conditional}
\title{Compare MID Conditional Expectations across Multiple Models with ggplot2}
\usage{
\method{ggmid}{midlist.conditional}(object, ...)

\method{autoplot}{midlist.conditional}(object, ...)
}
\arguments{
\item{object}{a "mid.conditional" object to be visualized.}

\item{...}{optional parameters passed on to the main layer.}
}
\value{
\code{ggmid.midlist.conditional()} returns a "ggplot" object.
}
\description{
For "mid.conditional" objects, \code{ggmid()} visualizes Individual Conditional Expectation (ICE) curves derived from a fitted MID model.
}
\details{
This is an S3 method for the \code{ggmid()} generic that produces ICE curves from a "mid.conditional" object.
ICE plots are a model-agnostic tool for visualizing how a model's prediction for a single observation changes as one feature varies.
This function plots one line for each observation in the data.

The \code{type} argument controls the visualization style:
The default, \code{type = "iceplot"}, plots the raw ICE curves.
The \code{type = "centered"} option creates the centered ICE (c-ICE) plot, where each curve is shifted to start at zero, making it easier to compare the slopes of the curves.

The \code{var.color}, \code{var.alpha}, etc., arguments allow you to map aesthetics to other variables in your data using (possibly) unquoted expressions.
}
\seealso{
\code{\link{ggmid.mid.conditional}}, \code{\link{plot.midlist.conditional}}
}
