% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_midlist.R
\name{labels.midrib}
\alias{labels.midrib}
\alias{labels<-}
\alias{labels<-.midrib}
\alias{labels<-.midlist}
\title{Modify Labels of a Collection of MID Models}
\usage{
\method{labels}{midrib}(object, ...)

labels(object) <- value

\method{labels}{midrib}(object) <- value

\method{labels}{midlist}(object) <- value
}
\arguments{
\item{object}{a "midrib" object, or its summary objects (e.g., "midlist.importance").}

\item{...}{optional parameters passed to other methods.}

\item{value}{a character vector of the same length as the number of models in the "midlist" object.}
}
\value{
\itemize{
  \item \code{labels} returns a character vector of model names.
  \item \code{labels<-} returns the updated object with new labels.
}
}
\description{
S3 methods to get/set the labels (model names) of a "midrib" or "midlist" object.
}
\details{
The "midrib" object stores multiple fitted MID models in a optimized struct-of-arrays format.

Because of the internal struct-of-arrays ("AsIs") structure, using \code{names()} on a "midrib" object returns internal component names (e.g., "intercept", "main.effects").
To safely access or modify the names of the models, always use \code{labels()} and \code{labels<-}.
}
