% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_conditional.R
\name{mid.conditional}
\alias{mid.conditional}
\alias{print.mid.conditional}
\title{Calculate ICE of MID Models}
\usage{
mid.conditional(
  object,
  variable,
  data,
  keep.effects = TRUE,
  n.samples = 100L,
  max.nrow = 100000L,
  type = c("response", "link")
)

\method{print}{mid.conditional}(x, ...)
}
\arguments{
\item{object}{a 'mid' object.}

\item{variable}{a character or expression specifying the variable for the ICE calculation.}

\item{data}{a data frame of observations.}

\item{keep.effects}{logical. If \code{TRUE}, the effects of component functions are stored in the output object.}

\item{n.samples}{an integer specifying the number of the sample points.}

\item{max.nrow}{an integer specifying the maximum number of rows of the output data frame.}

\item{type}{the type of the prediction. The default is "response". "link" is possible if the MID model uses a link function.}

\item{x}{a 'mid.conditional' object to print.}

\item{...}{additional parameters to be passed to \code{print.default()} to print sample points.}
}
\value{
\code{mid.conditional()} returns a 'mid.conditional' object with the following components:
\item{terms}{a character vector of relevant terms.}
\item{observed}{a data frame of the actual observations and the corresponding predictions.}
\item{conditional}{a data frame of the hypothetical observations and the corresponding predictions.}
\item{values}{a numeric vector of the sample points for the target variable.}
}
\description{
\code{mid.conditional()} creates an object to draw ICE curves of a MID model.
}
\details{
\code{mid.conditional()} obtains predictions for hypothetical observations from a 'mid' object and returns a 'mid.conditional' object.
The graphing functions \code{ggmid()} and \code{plot()} can be used to generate the ICE curve plots.
}
\examples{
data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
mc <- mid.conditional(mid, "Wind", airquality)
mc
}
