% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_extract.R
\name{mid.extract}
\alias{mid.extract}
\alias{mid.encoding.info}
\alias{mid.frames}
\alias{mid.terms}
\alias{terms.mid}
\alias{terms.mid.importance}
\title{Extract Information from a mid Object}
\usage{
mid.extract(object, component, ...)

mid.encoding.info(object, ...)

mid.frames(object, ...)

mid.terms(
  object,
  main.effect = TRUE,
  interaction = TRUE,
  require = NULL,
  remove = NULL,
  ...
)

\method{terms}{mid}(x, ...)

\method{terms}{mid.importance}(x, ...)
}
\arguments{
\item{object}{a mid object.}

\item{component}{a literal character string that specifies the name of the component to extract. Any component of the mid object can be extracted. In addition, some special values such as "frames", "encoding.info", and names of component functions like "x1:x2" can be used.}

\item{...}{optional arguments to be passed to the specific functions used to extract information.}

\item{main.effect}{logical. If FALSE, all main effects are excluded.}

\item{interaction}{logical. If FALSE, all interactions are excluded.}

\item{require}{a character vector of feature names. Only terms related to at least one of the specified features will be returned.}

\item{remove}{a character vector of feature names. All terms related to at least one of the specified features will not be returned.}

\item{x}{a mid object or a mid.importance object.}
}
\value{
\code{mid.extract()} returns the 'component' extracted from the mid object.
\code{mid.encoding.info()} returns basic information of the encoders.
\code{mid.frames()} returns encoding frames of each feature.
\code{mid.terms()} returns a character vector of term names.
}
\description{
Returns information on the mid object including a summary for encoders, uninterpreted rates, or other special components.
}
\examples{
data(trees, package = "datasets")
mid <- interpret(Volume ~ ., trees, k = 10)
mid.extract(mid, encoding.info)
mid.extract(mid, uninterpreted.rate)
mid.extract(mid, frames)
mid.extract(mid, Girth)
mid.extract(mid, intercept)
}
