% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_importance.R
\name{mid.importance}
\alias{mid.importance}
\alias{autoplot.mid.importance}
\alias{barplot.mid.importance}
\alias{plot.mid.importance}
\title{Calculate and Visualize MID-based Importance and Breakdown}
\usage{
mid.importance(x, data = NULL, weights = NULL, sort = TRUE, measure = 1L)

\method{autoplot}{mid.importance}(
  object,
  type = c("barplot", "heatmap"),
  max.terms = NA,
  scale.palette = c("#FFFFFF", "#464646"),
  ...
)

\method{barplot}{mid.importance}(object, max.terms = NA, ...)

\method{plot}{mid.importance}(
  x,
  type = c("barplot", "heatmap"),
  max.terms = NA,
  scale.palette = c("#FFFFFF", "#464646"),
  ...
)
}
\arguments{
\item{x}{a mid object.}

\item{data}{data to be used to calculate the importance. If NULL, the fitted matrix is extracted from the mid object. If the number of row equals to one, the mid breakdown will be calculated.}

\item{weights}{a numeric vector of weights.}

\item{sort}{logical. If TRUE, the data.frame will be sorted by magnitude of importance}

\item{measure}{an integer specifying the type of function to evaluate the importance of each effect. Possible values are "1" for the mean absolute effect, "2" for the root mean square effect, and "3" for the median absolute effect.}

\item{object}{a mid.importance object to be visualized.}

\item{type}{a character or an integer, specifying the type of the plot. Possible alternatives are 1 : "barplot" and 2 : "heatmap".}

\item{max.terms}{an integer, specifying the maximum number of functional decomposition terms to be plotted.}

\item{scale.palette}{color palette used to draw the interaction heatmap.}

\item{...}{optional arguments to be passed to graphic functions.}
}
\description{
Creates a data frame showing the importance of each functional decomposition term.
\code{mid.importance()} returns a object of class "mid.importance", for which methods for \code{ggplot2::autoplot()} and \code{graphics::barplot()} are defined.
}
\examples{
data(airquality, package = "datasets")
model <- glm(Ozone ~ Solar.R + Temp + Wind + Month, "poisson", airquality)
mid <- interpret(Ozone ~ Solar.R + Temp + Wind + Month, airquality, model)
mid.importance(mid)
ggplot2::autoplot(mid.importance(mid))
}
