% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_plots.R
\name{mid.plots}
\alias{mid.plots}
\title{Create Plots for Multiple Functional Decomposition Terms}
\usage{
mid.plots(
  x,
  terms = mid.terms(x, main.effect = T, interaction = F),
  limits = c(NA, NA),
  add.intercept = FALSE,
  max.plots = 16L,
  engine = c("ggplot2", "base", "graphics"),
  ...
)
}
\arguments{
\item{x}{a mid object.}

\item{terms}{a character vector that specifies the names of the terms to be visualized.}

\item{limits}{NULL or a numeric vector of length two providing limits of the scale. NA will be replaced by the minimum or maximum mid value in all terms.}

\item{add.intercept}{logical. If TRUE, the intercept is added to the mid values and the scale for the plot is shifted.}

\item{max.plots}{the number of maximum number of plots.}

\item{engine}{a name of the package used to create plots. Possible values are "ggplot2" and "base" ("graphics").}

\item{...}{optional parameters to be passed to ggmid() function.}
}
\description{
Returns a list of ggplot objects, each representing the mid values for the corresponding functional decomposition term.
}
\examples{
data(diamonds, package = "ggplot2")
diamonds <- diamonds[sample(nrow(diamonds), 1e4L), ]
mid <- interpret(price ~ (carat + cut + color + clarity) ^ 2, diamonds)
mid.plots(mid, c("carat", "color", "carat:color", "clarity:color"), NULL)
}
