% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midr_options.R
\name{midr.options}
\alias{midr.options}
\alias{midr.par}
\title{Set or Get Global Options for the midr Package}
\usage{
midr.options(...)

midr.par(...)
}
\arguments{
\item{...}{For \code{midr.options()}: options to be defined in \code{name = value} form.
If no arguments are provided, it returns all current package options.
If character strings are provided, it returns the values of those options.
For \code{midr.par()}: graphical parameters in \code{name = value} form can be supplied as arguments.}
}
\value{
When called without arguments, \code{midr.options()} returns a named list of all current options with the \code{midr.} prefix.
When called with character strings, \code{midr.options()} returns a named list of the requested options.
When called with \code{name = value} pairs, \code{midr.options()} invisibly returns a named list of the previous values.

\code{midr.par()} returns the previous values of the changed parameters in an invisible named list.
}
\description{
\code{midr.options()} manages global settings specific to the \strong{midr} package, such as computational solvers and color theme environments.

\code{midr.par()} manages graphical parameters for base R graphics.
}
\details{
\code{midr.options()} provides an interface to R's global \code{options()} but safely scopes all parameters with a \code{midr.} prefix.
 To prevent typos, it validates option names against expected prefixes:
\itemize{
  \item \code{solver.*}: For registering custom least-squares solvers; e.g., \code{solver.TAG = FUN}, where \code{FUN(x, y)} returns a list containing at least \code{coefficients} and possibly \code{residuals} and \code{rank}. The custom solver can be called via \code{interpret(method = "TAG")}.
  \item \code{color.theme.*}: For configuring color theme settings and registries.
  \item \code{verbosity}: For controlling message verbosity. See \code{\link{interpret}()}.
}

\code{midr.par()} wraps \code{\link[graphics]{par}()} to enforce a themed aesthetic as the default, which can be further customized by the user.
}
