% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid_breakdown.R
\name{plot.mid.breakdown}
\alias{plot.mid.breakdown}
\title{Plot MID Breakdown with Basic Functions}
\usage{
\method{plot}{mid.breakdown}(
  x,
  type = c("barplot", "heatmap", "dotchart", "boxplot"),
  max.bars = 15L,
  vline = TRUE,
  sep = " = ",
  terms = NULL,
  catchall = "others",
  ...
)
}
\arguments{
\item{x}{a "mid.breakdown" object to be visualized.}

\item{type}{a character string specifying the type of the plot. One of "barplot" or "dotchart".}

\item{max.bars}{an integer specifying the maximum number of bars in the barplot, boxplot and dotchart.}

\item{vline}{logical. If \code{TRUE}, the vertical line is drawn at zero or the intercept.}

\item{sep}{a character string to separate terms and values. Default is the return.}

\item{terms}{an optional character vector specifying the terms to be displayed.}

\item{catchall}{a character string to be used as the catchall label.}

\item{...}{optional parameters to be passed to the graphing function.}
}
\value{
\code{plot.mid.breakdown()} produces a plot and returns \code{NULL}.
}
\description{
For "mid.breakdown" objects, \code{plot()} visualizes the breakdown of a prediction by component functions.
}
\details{
The S3 method of \code{plot()} for "mid.breakdown" objects creates a visualization of the MID breakdown using \code{graphics::barplot()} or \code{graphics::dotchart()}.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
mbd <- mid.breakdown(mid, diamonds[1L, ])
plot(mbd)
plot(mbd, type = "dotchart")
}
