% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid_importance.R
\name{plot.mid.importance}
\alias{plot.mid.importance}
\title{Plot Term Importance with graphics Package}
\usage{
\method{plot}{mid.importance}(
  x,
  type = c("barplot", "heatmap"),
  max.terms = NA,
  scale.palette = c("#FFFFFF", "#464646"),
  ...
)
}
\arguments{
\item{x}{a mid.importance object to plot.}

\item{type}{a character or an integer, specifying the type of the plot. Possible alternatives are "barplot" and "heatmap".}

\item{max.terms}{an integer, specifying the maximum number of component terms to be plotted in the barplot.}

\item{scale.palette}{color palette used to draw the interaction heatmap.}

\item{...}{optional arguments to be passed to graphic functions.}
}
\description{
Creates a plot showing the MID-based term importance
}
