% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mid.R
\name{print.mid}
\alias{print.mid}
\title{Print Method for MID Objects}
\usage{
\method{print}{mid}(x, digits = max(3L, getOption("digits") - 2L), omit.values = FALSE, ...)
}
\arguments{
\item{x}{a mid object to be printed.}

\item{digits}{number of significant digits.}

\item{omit.values}{logical. If TRUE, mid values of main effect terms are not printed.}

\item{...}{not used.}
}
\value{
\code{print.mid()} returns the 'mid' object passed to the function without any modification.
}
\description{
Prints the summary of the fitted mid object.
}
\examples{
data(cars, package = "datasets")
print(interpret(dist ~ speed, cars))
}
