% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_mid.R
\name{shapviz.mid}
\alias{shapviz.mid}
\title{Calculate MID-Derived Shapley Values}
\usage{
\method{shapviz}{mid}(object, data = NULL)
}
\arguments{
\item{object}{a "mid" object.}

\item{data}{a data frame containing the observations for which to calculate MID-derived Shapley values. If not passed, data is automatically extracted based on the function call.}
}
\value{
\code{shapviz.mid()} returns an object of class "shapviz".
}
\description{
\code{shapviz.mid()} is an S3 method for the \code{shapviz::shapviz()} generic, which calculates MID-derived Shapley values from a fitted MID model.
}
\details{
The function calculates MID-derived Shapley values by attributing the contribution of each component function to its respective variables as follows:
first, each main effect is fully attributed to its corresponding variable; and
then, each second-order interaction effect is split equally between the two variables involved.
}
