% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_effect.R
\name{term.effect}
\alias{term.effect}
\alias{term.effect.mid}
\alias{term.effect.mids}
\alias{term.effect.default}
\alias{term.f}
\title{Evaluate Single Component Functions of Additive Models}
\usage{
term.effect(object, ...)

\method{term.effect}{mid}(object, term, x, y = NULL, ...)

\method{term.effect}{mids}(object, term, x, y = NULL, ...)

\method{term.effect}{default}(object, term, x, y = NULL, data = NULL, ...)

term.f(object, ...)
}
\arguments{
\item{object}{a "mid" object, a collection of models ("mids"), or other compatible additive model objects (e.g., "lm", "glm").}

\item{...}{optional arguments to be passed to methods.}

\item{term}{a character string specifying the component function (term) to evaluate.}

\item{x}{a vector of values for the first variable in the term. If a matrix or data frame is provided, values of the related variables are automatically extracted from it.}

\item{y}{a vector of values for the second variable in an interaction term. Ignored if \code{x} is a data frame containing both variables.}

\item{data}{an optional data frame containing the original training data. Required for additive models if the data cannot be automatically extracted from the model object.}
}
\value{
\code{term.effect()} returns a numeric vector of the calculated term contributions, with the same length as \code{x}.

For a collection of models ("mids"), \code{term.effect()} returns a numeric matrix where each column corresponds to a model.
}
\description{
\code{term.effect()} calculates the contribution of a single component function of a fitted additive models.
It serves as a low-level helper function for making predictions or for direct analysis of a term effect.

\code{term.f()} is a convenient shorthand for \code{term.effect()}.
}
\details{
\code{term.effect()} is a low-level S3 generic function designed to calculate the contribution of a single component function.
Unlike \code{predict.mid()}, which is designed to return total model predictions, \code{term.effect()} is more flexible.
It accepts vectors, as well as matrices or data frames, as input for \code{x} and \code{y}. If \code{x} is a data frame, the necessary columns are automatically extracted.
This makes it particularly useful for visualizing a component's effect in combination with standard plotting functions, such as \code{graphics::curve()}.

For a main effect, the function evaluates the component function \eqn{f_j(x_j)} for a vector of values \eqn{x_j}.
For an interaction, it evaluates \eqn{f_{jk}(x_j, x_k)} using vectors \eqn{x_j} and \eqn{x_k}.
The assignment of \code{x} and \code{y} strictly follows the order of variables specified in the \code{term} argument.
For example, if \code{term = "Temp:Wind"}, \code{x} is assigned to \code{Temp} and \code{y} is assigned to \code{Wind}.
}
\examples{
data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, data = airquality, lambda = 1)

# Visualize the main effect of "Wind"
curve(term.effect(mid, term = "Wind", x), from = 0, to = 25)

# Visualize the interaction of "Wind" and "Temp"
curve(term.f(mid, "Wind:Temp", x, 50), 0, 25)
curve(term.f(mid, "Wind:Temp", x, 60), 0, 25, add = TRUE, lty = 2)
curve(term.f(mid, "Wind:Temp", x, 70), 0, 25, add = TRUE, lty = 3)
}
