% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{weighted.quantile}
\alias{weighted.quantile}
\title{Weighted Sample Quantile}
\usage{
weighted.quantile(
  x,
  w = NULL,
  probs = seq(0, 1, 0.25),
  na.rm = FALSE,
  names = TRUE,
  digits = 7L,
  type = 1L,
  ...
)
}
\arguments{
\item{x}{an object containing the values whose weighted quantiles is to be computed.}

\item{w}{a numeric vector of the same length as 'x' giving the weights to use for elements of it.}

\item{probs}{a numeric vector of probabilities with values in [0, 1].}

\item{na.rm}{logical. If TRUE, any \code{NA} and \code{NaN}s are removed from 'x' before the quantiles are computed.}

\item{names}{logical. If TRUE, the result has a names attribute.}

\item{digits}{used only when \code{names} is TRUE. The precision to use when formatting the percentages.}

\item{type}{an integer selecting the quantile algorithms. Only 1 is available for the weighted quantile.}

\item{...}{further arguments passed to \code{stats::quantile()} when the weights is not passed.}
}
\description{
Returns weighted sample quantiles corresponding to the given probabilities. For the weighted quantiles, only "type 1" quantiles of \code{stats::quantile()} (the inverse of empirical distribution function) is available.
}
