% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted}
\alias{weighted}
\alias{augmented}
\alias{shuffled}
\alias{latticized}
\alias{weights.weighted}
\title{Weighted Data Frames}
\usage{
weighted(data, weights = NULL, ...)

augmented(data, weights = NULL, size = nrow(data), r = 0.01)

shuffled(data, weights = NULL, size = nrow(data))

latticized(
  data,
  weights = NULL,
  k = 10L,
  type = 1L,
  use.catchall = TRUE,
  catchall = "(others)",
  frames = list(),
  keep.mean = TRUE
)

\method{weights}{weighted}(object, ...)
}
\arguments{
\item{data}{a data frame.}

\item{weights}{a numeric vector specifying the sample weights for each observation.}

\item{...}{not used.}

\item{size}{an integer specifying the number of random observations to add to the data frame.}

\item{r}{a double. The weight for the random observations is calculated as \code{sum(attr(data, "weights")) * r / size}.}

\item{k}{an integer or a numeric vector of length two, specifying the maximum number of sample points for each variable. If an integer is passed, \code{k} is used for main effects and \code{sqrt(k)} is used for interactions. If not positive, all unique values are used as sample points.}

\item{type}{an integer or a numeric vector of length two, specifying the type of encoding.}

\item{use.catchall}{logical. If \code{TRUE}, less frequent levels of factor variables are dropped and replaced by the catchall level.}

\item{catchall}{a character specifying the catchall level.}

\item{frames}{a named list of encoding frames.}

\item{keep.mean}{logical. If \code{TRUE}, the representative values of each group is the mean of the corresponding group.}

\item{object}{a weighted data frame.}
}
\value{
\code{weighted()} returns a data frame with the attribute 'weights'.
\code{augmented()} returns a weighted data frame of the original data and the shuffled data with relatively small weights.
\code{shuffled()} returns a weighted data frame of the shuffled data.
\code{latticized()} returns a weighted data frame of latticized data, whose values are grouped and replaced by the representative value of the corresponding group.
}
\description{
\code{weighted()} returns a data frame with sample weights.
}
\details{
\code{weighted()} returns a data frame with the 'weights' attribute that can be extracted using \code{stats::weights()}, and three derivative functions, \code{augmented()}, \code{shuffled()} and \code{latticized()}, return a modified data frame with sample weights.
These functions are designed for use with \code{interpret()}.
Since the modified data frames do not preserve the original correlation structure between the variables, the values of the objective variable (y) should always be replaced by the model predictions (yhat).
}
\examples{
x1 <- runif(1000L, -1, 1)
x2 <- x1 + runif(1000L, -1, 1)
X <- weighted(cbind(x1, x2), (abs(x1) + abs(x2)) / 2)
Y1 <- weighted(X)
ggplot2::ggplot(Y1, ggplot2::aes(x1, x2, alpha = weights(Y1))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "weighted") +
  ggplot2::theme_bw()
Y2 <- shuffled(X)
ggplot2::ggplot(Y2, ggplot2::aes(x1, x2, alpha = weights(Y2))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "shuffled") +
  ggplot2::theme_bw()
Y3 <- augmented(X)
ggplot2::ggplot(Y3, ggplot2::aes(x1, x2, alpha = weights(Y3))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "augmented") +
  ggplot2::theme_bw()
Y4 <- latticized(X)
ggplot2::ggplot(Y4, ggplot2::aes(x1, x2, size = weights(Y4))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "latticized") +
  ggplot2::theme_bw()
}
