% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{weighted.tabulate}
\alias{weighted.tabulate}
\title{Weighted Tabulation for Vectors}
\usage{
weighted.tabulate(bin, w = NULL, nbins = max(1L, bin, na.rm = TRUE))
}
\arguments{
\item{bin}{a numeric vector of positive integers or a factor.}

\item{w}{a numerical vector of the same length as 'bin' giving the weights to use for elements of it.}

\item{nbins}{the number of bins to be used.}
}
\description{
Returns the sum of weights for each interger occurs in the integer valued argument \code{bin}.
}
