% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_midlist.R
\name{extract.midlist}
\alias{extract.midlist}
\alias{[.midrib}
\alias{[[.midrib}
\alias{[.midlist}
\title{Subset MID Models in a Collection}
\usage{
\method{[}{midrib}(x, i, drop = if (missing(i)) TRUE else length(i) == 1L)

\method{[[}{midrib}(x, i, exact = TRUE)

\method{[}{midlist}(x, i, drop = if (missing(i)) TRUE else length(i) == 1L)
}
\arguments{
\item{x}{a "midrib" object, a "midlist" object, or its summary objects (e.g., "midlist.importance").}

\item{i}{indices specifying elements to extract. Can be numeric, character, or logical vectors.}

\item{drop}{logical. If \code{TRUE} the result is coerced to the lowest possible dimension. For a "midrib", extracting a single element drops it to a "mid" object.}

\item{exact}{logical. If \code{TRUE}, exact matching is used for character strings.}
}
\value{
\code{[} returns a "midrib", "midlist" or "mid" object.

\code{[[} returns a single "mid" object.
}
\description{
S3 methods to extract parts of a "midrib" or "midlist" object.
}
\details{
The "midrib" object stores multiple fitted MID models in a optimized struct-of-arrays format.

When extracting multiple models using \code{[}, it returns a subsetted "midrib" or "midlist" object.
By default, if a single model is extracted (\code{length(i) == 1L}) and \code{drop = TRUE}, the object is simplified to a single "mid" object.
\code{[[} always extracts a single model as a "mid" object.

Similar extraction rules apply to summary objects like "midlist.importance".
}
