% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_encoder.R
\name{factor.encoder}
\alias{factor.encoder}
\alias{factor.frame}
\title{Encoder for Qualitative Variables}
\usage{
factor.encoder(
  x,
  k,
  use.catchall = TRUE,
  catchall = "(others)",
  tag = "x",
  frame = NULL,
  weights = NULL
)

factor.frame(levels, catchall = "(others)", tag = "x")
}
\arguments{
\item{x}{a vector to be encoded as a qualitative variable.}

\item{k}{an integer specifying the max number of the levels after the encoding. If not positive, all unique values of x are chosen as sample points.}

\item{use.catchall}{logical. If TRUE, less frequent levels are dropped and replaced with the catchall level.}

\item{catchall}{a character used as the name of "catchall" level for unused levels.}

\item{tag}{name of the variable.}

\item{frame}{a \code{factor.frame} object or a vector, containing the information about the levels of the variable.}

\item{weights}{optional. a numeric vector indicating the weight of each value of 'x'.}

\item{levels}{a vector to be used as the levels of the variable.}
}
\value{
\code{factor.encoder()} returns a list containing the following components:
\item{frame}{a data frame containing the encoding information.}
\item{encode}{a function to encode new data into a dummy matrix.}
\item{n}{the number of encoding levels.}
\item{type}{type of encoding.}
}
\description{
Returns a list consisting of the encoding information of a predictor variable as a qualitative variable.
}
\examples{
data(iris, package = "datasets")
enc <- factor.encoder(x = iris$Species, use.catchall = FALSE)
enc$frame
enc$encode(new_x = c("setosa", "virginica", "ensata", NA, "versicolor"))
}
