% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yhat.R
\name{get.yhat}
\alias{get.yhat}
\alias{get.yhat.default}
\alias{get.yhat.mid}
\alias{get.yhat.lm}
\alias{get.yhat.glm}
\alias{get.yhat.rpart}
\alias{get.yhat.randomForest}
\alias{get.yhat.ranger}
\alias{get.yhat.svm}
\alias{get.yhat.ksvm}
\alias{get.yhat.AccurateGLM}
\alias{get.yhat.model_fit}
\title{Get Predicted Values from Fitted Models}
\usage{
get.yhat(X.model, newdata, ...)

\method{get.yhat}{default}(X.model, newdata, ...)

\method{get.yhat}{mid}(X.model, newdata, ...)

\method{get.yhat}{lm}(X.model, newdata, ...)

\method{get.yhat}{glm}(X.model, newdata, ...)

\method{get.yhat}{rpart}(X.model, newdata, ...)

\method{get.yhat}{randomForest}(X.model, newdata, ...)

\method{get.yhat}{ranger}(X.model, newdata, ...)

\method{get.yhat}{svm}(X.model, newdata, ...)

\method{get.yhat}{ksvm}(X.model, newdata, ...)

\method{get.yhat}{AccurateGLM}(X.model, newdata, ...)

\method{get.yhat}{model_fit}(X.model, newdata, ...)
}
\arguments{
\item{X.model}{a model object to be interpreted.}

\item{newdata}{a data.frame or a matrix.}

\item{...}{other parameters that will be passed to the predict function.}
}
\value{
\code{get.yhat()} returns a numeric vector of model predictions.
}
\description{
Returns the predicted values or the target class probabilities of a predictive model for the \code{newdata}.
}
\examples{
data(trees, package = "datasets")
model <- glm(Volume ~ ., trees, family = Gamma(log))
predict(model, trees, "response")
get.yhat(model, trees)
}
