% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_breakdown.R
\name{ggmid.mid.breakdown}
\alias{ggmid.mid.breakdown}
\alias{autoplot.mid.breakdown}
\title{Plot MID Breakdown with ggplot2 Package}
\usage{
\method{ggmid}{mid.breakdown}(
  object,
  type = c("barplot", "dotchart", "waterfall"),
  plot.main = TRUE,
  max.bars = NA,
  sep = "\\n",
  width = 0.8,
  ...
)

\method{autoplot}{mid.breakdown}(object, ...)
}
\arguments{
\item{object}{a "mid.breakdown" object to be visualized.}

\item{type}{a character string specifying the type of the plot. One of "barplot", "dotchart" or "waterfall".}

\item{plot.main}{logical. If \code{TRUE}, the main layer is not drawn.}

\item{max.bars}{an integer specifying the maximum number of bars in the plot.}

\item{sep}{a character string to separate terms and values. Default is the return.}

\item{width}{a numeric value specifying the width of the bars.}

\item{...}{optional parameters to be passed to the main layer.}
}
\value{
\code{ggmid.mid.breakdown()} returns a "ggplot" object.
}
\description{
For "mid.breakdown" objects, \code{ggmid()} visualizes the breakdown of a prediction by component functions.
}
\details{
The S3 method of \code{ggmid()} for "mid.breakdown" objects creates a "ggplot" object that visualizes the breakdown of a single model prediction.
The main layer is drawn using \code{geom_col()}.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
mbd <- mid.breakdown(mid, diamonds[1L, ])
ggmid(mbd)
ggmid(mbd, type = "dotchart")
ggmid(mbd, type = "waterfall")
}
