% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_conditional.R
\name{ggmid.mid.conditional}
\alias{ggmid.mid.conditional}
\alias{autoplot.mid.conditional}
\title{Plot ICE of MID Model with ggplot2 Package}
\usage{
\method{ggmid}{mid.conditional}(
  object,
  limits = c(NA, NA),
  plot.main = TRUE,
  centered = FALSE,
  draw.dots = TRUE,
  sample = NULL,
  term = NULL,
  variable.alpha = NULL,
  variable.colour = NULL,
  variable.linetype = NULL,
  variable.linewidth = NULL,
  ...
)

\method{autoplot}{mid.conditional}(object, ...)
}
\arguments{
\item{object}{a "mid.conditional" object to be visualized.}

\item{limits}{\code{NULL} or a numeric vector of length two specifying the limits of the scale. \code{NA}s are replaced by the minimum and/or maximum MID values.}

\item{plot.main}{logical. If \code{FALSE}, the main layer is not drawn.}

\item{centered}{logical. If \code{TRUE}, the ICE values of each observation are set to zero at the leftmost point of the variable.}

\item{draw.dots}{logical. If \code{TRUE}, the points representing the predictions for each observation are plotted.}

\item{sample}{an optional vector specifying the names of observations to be plotted.}

\item{term}{an optional character string specifying an interaction term. If passed, the ICE curve for the specified term is plotted.}

\item{variable.alpha}{a name of the variable to be used to set \code{alpha}.}

\item{variable.colour}{a name of the variable to be used to set \code{colour}.}

\item{variable.linetype}{a name of the variable to be used to set \code{linetype}.}

\item{variable.linewidth}{a name of the variable to be used to set \code{linewidth}.}

\item{...}{optional parameters to be passed to the main layer.}
}
\value{
\code{ggmid.mid.conditional()} returns a "ggplot" object.
}
\description{
For "mid.conditional" objects, \code{ggmid()} visualizes ICE curves of a MID model.
}
\details{
The S3 method of \code{ggmid()} for "mid.conditional" objects creates a "ggplot" object that visualizes ICE curves of a fitted MID model using \code{geom_line()}.
}
\examples{
data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
mc <- mid.conditional(mid, "Wind", airquality)
ggmid(mc, variable.colour = "Solar.R", centered = TRUE)
}
