% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_conditional.R
\name{ggmid.mid.conditional}
\alias{ggmid.mid.conditional}
\alias{autoplot.mid.conditional}
\title{Plot MID Individual Conditional Expectations with ggplot2 Package}
\usage{
\method{ggmid}{mid.conditional}(
  object,
  limits = c(NA, NA),
  plot.main = TRUE,
  centered = FALSE,
  draw.dots = TRUE,
  sample = NULL,
  term = NULL,
  variable.alpha = NULL,
  variable.colour = NULL,
  variable.linetype = NULL,
  variable.linewidth = NULL,
  ...
)

\method{autoplot}{mid.conditional}(object, ...)
}
\arguments{
\item{object}{a mid.conditional object to visualize.}

\item{limits}{NULL or a numeric vector of length two providing limits of the scale. NA is replaced by the minimum or maximum mid value.}

\item{plot.main}{logical. If TRUE, lines representing the individual conditional expectations are drawn.}

\item{centered}{logical.}

\item{draw.dots}{logical. If TRUE, points representing the predictions at the observed values are}

\item{sample}{a vector specifying the set of names of the observations to be plotted.}

\item{term}{an optional character specifying one of the relevant terms. If passed, the individual conditional expectations for the specified term are plotted.}

\item{variable.alpha}{a name of the predictor variable to use to set \code{alpha} for each plot.}

\item{variable.colour}{a name of the predictor variable to use to set \code{color} for each plot.}

\item{variable.linetype}{a name of the predictor variable to use to set \code{linetype} for each plot.}

\item{variable.linewidth}{a name of the predictor variable to use to set \code{linewidth} for each plot.}

\item{...}{optional parameters to be directly passed to \code{ggplot2::geom_line()}.}
}
\description{
Creates a ggplot object representing the MID-based individual conditional expectations
}
