% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_importance.R
\name{ggmid.mid.importance}
\alias{ggmid.mid.importance}
\alias{autoplot.mid.importance}
\title{Plot Term Importance with ggplot2 Package}
\usage{
\method{ggmid}{mid.importance}(
  object,
  type = c("barplot", "heatmap"),
  plot.main = TRUE,
  max.terms = NA,
  scale.palette = c("#FFFFFF", "#464646"),
  ...
)

\method{autoplot}{mid.importance}(object, ...)
}
\arguments{
\item{object}{a mid.importance object to be visualized.}

\item{type}{a character or an integer, specifying the type of the plot. Possible alternatives are "barplot" and "heatmap".}

\item{plot.main}{logical. If TRUE, lines, bars or rectangles representing mid values are drawn.}

\item{max.terms}{an integer, specifying the maximum number of component terms to be plotted in the barplot.}

\item{scale.palette}{color palette used to draw the interaction heatmap.}

\item{...}{optional arguments to be passed to graphic functions.}
}
\description{
Creates a ggplot object representing the MID-based term importance
}
