% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid_mid_importance.R
\name{ggmid.mid.importance}
\alias{ggmid.mid.importance}
\alias{autoplot.mid.importance}
\title{Plot MID Importance with ggplot2}
\usage{
\method{ggmid}{mid.importance}(
  object,
  type = c("barplot", "dotchart", "heatmap", "boxplot"),
  theme = NULL,
  max.bars = 30L,
  ...
)

\method{autoplot}{mid.importance}(object, ...)
}
\arguments{
\item{object}{a "mid.importance" object to be visualized.}

\item{type}{a character string specifying the type of the plot. One of "barplot", "heatmap", "dotchart" or "boxplot".}

\item{theme}{a character string specifying the color theme or any item that can be used to define "color.theme" object.}

\item{max.bars}{an integer specifying the maximum number of bars in the barplot, boxplot and dotchart.}

\item{...}{optional parameters to be passed to the main layer.}
}
\value{
\code{ggmid.mid.importance()} returns a "ggplot" object.
}
\description{
For "mid.importance" objects, \code{ggmid()} visualizes the importance of component functions of the fitted MID model.
}
\details{
This is an S3 method for the \code{ggmid()} generic that creates an importance plot from a "mid.importance" object, visualizing the average contribution of component functions to the fitted MID model.

The \code{type} argument controls the visualization style.
The default, \code{type = "barplot"}, creates a standard bar plot where the length of each bar represents the overall importance of the term.
The \code{type = "dotchart"} option creates a dot plot, offering a clean alternative to the bar plot for visualizing term importance.
The \code{type = "heatmap"} option creates a matrix-shaped heat map where the color of each cell represents the importance of the interaction between a pair of variables, or the main effect on the diagonal.
The \code{type = "boxplot"} option creates a box plot where each box shows the distribution of a term's contributions across all observations, providing insight into the varibability of each term's effect.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
imp <- mid.importance(mid)
ggmid(imp, theme = "Tableau 10")
ggmid(imp, type = "dotchart", theme = "Okabe-Ito", size = 3)
ggmid(imp, type = "heatmap", theme = "Blues")
ggmid(imp, type = "boxplot", theme = "Accent")
}
\seealso{
\code{\link{ggmid}}, \code{\link{plot.mid.importance}}
}
