% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid.R
\name{ggmid}
\alias{ggmid}
\alias{ggmid.mid}
\alias{autoplot.mid}
\title{Plot MID with ggplot2 Package}
\usage{
ggmid(object, ...)

\method{ggmid}{mid}(
  object,
  term,
  limits = c(NA, NA),
  plot.main = TRUE,
  add.intercept = FALSE,
  include.main.effects = FALSE,
  interaction.type = c("default", "raster", "rectangle"),
  scale.type = "default",
  scale.palette = c("#2f7a9a", "#FFFFFF", "#7e1952"),
  cells.count = c(100L, 100L),
  ...
)

\method{autoplot}{mid}(object, ...)
}
\arguments{
\item{object}{a "mid" object to be visualized.}

\item{...}{optional parameters to be passed to the main layer.}

\item{term}{a character string specifying the component function to be plotted.}

\item{limits}{\code{NULL} or a numeric vector of length two specifying the limits of the plotting scale. \code{NA}s are replaced by the minimum and/or maximum MID values.}

\item{plot.main}{logical. If \code{FALSE}, the main layer is not drawn.}

\item{add.intercept}{logical. If \code{TRUE}, the intercept is added to the MID values.}

\item{include.main.effects}{logical. If \code{TRUE}, the main effects are included in the interaction plot.}

\item{interaction.type}{character string. The method for plotting the interaction effects.}

\item{scale.type}{a character string or function specifying the color type of interaction plot. One of "default", "viridis", "gradient" or a function that returns a continuous color scale for \code{fill} aesthetics.}

\item{scale.palette}{a character vector of color names. The colors are used for the interaction plot when \code{scale.type} is "default".}

\item{cells.count}{an integer or integer-valued vector of length two, specifying the number of cells for the raster type interaction plot.}
}
\value{
\code{ggmid.mid()} returns a "ggplot" object.
}
\description{
For "mid" objects, \code{ggmid()} visualizes a MID component function using the ggplot2 package.
}
\details{
The S3 method of \code{ggmid()} for "mid" objects creates a "ggplot" object that visualizes a MID component function.
The main layer is drawn using \code{geom_line()} or \code{geom_path()} for a main effect of a quantitative variable, \code{geom_col()} for a main effect of a qualitative variable, and \code{geom_raster()} or \code{geom_rect()} for an interaction effect.
For other methods of \code{ggmid()}, see \code{help(ggmid.mid.importance)} or \code{help(ggmid.mid.conditional)}.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
ggmid(mid, "carat")
ggplot2::autoplot(mid, "clarity")
ggmid(mid, "carat:clarity")
ggmid(mid, "carat:clarity", add.intercept = TRUE,
      include.main.effects = TRUE, scale.type = "viridis")
}
