% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmid.R
\name{ggmid}
\alias{ggmid}
\alias{ggmid.mid}
\alias{autoplot.mid}
\title{Plot MID Component Functions with ggplot2}
\usage{
ggmid(object, ...)

\method{ggmid}{mid}(
  object,
  term,
  type = c("effect", "data", "compound"),
  theme = NULL,
  intercept = FALSE,
  main.effects = FALSE,
  data = NULL,
  limits = c(NA, NA),
  jitter = 0.3,
  cells.count = c(100L, 100L),
  ...
)

\method{autoplot}{mid}(object, ...)
}
\arguments{
\item{object}{a "mid" object to be visualized.}

\item{...}{optional parameters passed to the main plotting layer.}

\item{term}{a character string specifying the component function to be plotted.}

\item{type}{the method for plotting. One of "effect", "data" or "compound".}

\item{theme}{a character string or object defining the color theme.}

\item{intercept}{logical. If \code{TRUE}, the intercept is added to the MID values.}

\item{main.effects}{logical. If \code{TRUE}, main effects are included in the interaction plot.}

\item{data}{a data frame to be plotted with the corresponding MID values. If not provided, data is automatically extracted.}

\item{limits}{a numeric vector of length two specifying the limits of the plotting scale. \code{NA} values are replaced by the minimum and/or maximum MID values.}

\item{jitter}{a numeric value specifying the amount of jitter for the data points.}

\item{cells.count}{an integer or vector of two integers specifying the resolution of the raster plot for interactions.}
}
\value{
\code{ggmid.mid()} returns a "ggplot" object.
}
\description{
\code{ggmid()} is a plotting function for "mid" objects, creating a "ggplot" object to visualize a single component function.
}
\details{
The \code{type} argument controls the plotting mode.
When \code{type = "effect"} (default), the component function itself is plotted.
The main layer is automatically selected based on the effect's type:
for a quantitative main effect, a line plot is drawn with \code{geom_line()} or \code{geom_path()};
for a qualitative main effect, a bar plot is drawn with \code{geom_col()}; and
for an interaction effect, a raster plot or rectangle plot is drawn with \code{geom_raster()} or \code{geom_rect}.
If \code{type = "data"}, only the values of the component function at the raw data points are plotted, which requires the \code{data} argument to be supplied.
Finally, \code{type = "compound"} combines both modes, plotting the component function as well as the raw data points.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])

# Plot a quantitative main effect
ggmid(mid, "carat")

# Plot a qualitative main effect
ggmid(mid, "clarity")

# Plot an interaction effect with data points and a raster layer
ggmid(mid, "carat:clarity", type = "compound", data = diamonds[idx, ])

# Use a different colot theme
ggmid(mid, "clarity:color", theme = "RdBu")
}
\seealso{
\code{\link{interpret}}, \code{\link{ggmid.mid.importance}}, \code{\link{ggmid.mid.conditional}}, \code{\link{ggmid.mid.breakdown}}
}
