% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_breakdown.R
\name{mid.breakdown}
\alias{mid.breakdown}
\alias{print.mid.breakdown}
\title{Calculate MID Breakdown}
\usage{
mid.breakdown(
  object,
  data,
  sort = TRUE,
  digits = max(3L, getOption("digits") - 2L),
  left = "(",
  right = ")",
  sep = ", "
)

\method{print}{mid.breakdown}(x, digits = max(3L, getOption("digits") - 2L), ...)
}
\arguments{
\item{object}{a "mid" object.}

\item{data}{a data frame containing the observations to be used to calculate the MID importance. If \code{NULL}, the \code{fitted.matrix} of the MID model is used. If the \code{data} has only one observation, the output has the special class "mid.breakdown".}

\item{sort}{logical. If \code{TRUE}, the output data frame is sorted by MID importance.}

\item{digits}{an integer specifying the minimum number of significant digits to be printed.}

\item{left}{a character string to be used as the left bracket of the value.}

\item{right}{a character string to be used as the right bracket of the value.}

\item{sep}{a character string to separate the two values of interaction terms.}

\item{x}{a "mid.importance" object to be printed.}

\item{...}{additional parameters to be passed to \code{print.data.frame()} to print the importance of component functions.}
}
\value{
\code{mid.importance} returns an object of the class "mid.importance" containing the following components.
\item{importance}{the data frame of calculated importances.}
\item{predictions}{the matrix of the fitted or predicted MID values.}
\item{measure}{the type of the importance measure.}
}
\description{
\code{mid.breakdown()} calculates the MID breakdown of a prediction of the MID model.
}
\details{
\code{mid.breakdown()} returns an object of class "mid.breakdown".
}
\examples{
data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
mbd <- mid.breakdown(mid, airquality[1L, ])
mbd
}
