% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_extract.R
\name{mid.extract}
\alias{mid.extract}
\alias{mid.encoding.info}
\alias{mid.frames}
\alias{mid.terms}
\alias{terms.mid}
\title{Extract Information from a mid Object}
\usage{
mid.extract(x, component, ...)

mid.encoding.info(x, ...)

mid.frames(x, ...)

mid.terms(
  x,
  main.effect = TRUE,
  interaction = TRUE,
  require = NULL,
  remove = NULL,
  ...
)

\method{terms}{mid}(x, ...)
}
\arguments{
\item{x}{a mid object.}

\item{component}{a literal character string that specifies the name of the component to extract. Any component of the mid object can be extracted. In addition, some special values such as "importance", "UR", "encoders", and names of any functional decomposition terms like "x1:x2" can be used.}

\item{...}{optional arguments to be passed to the specific functions used to extract information.}

\item{main.effect}{logical. If FALSE, all main effects are excluded.}

\item{interaction}{logical. If FALSE, all interactions are excluded.}

\item{require}{a character vector of feature names. Only terms related to at least one of the specified features will be returned.}

\item{remove}{a character vector of feature names. All terms related to at least one of the specified features will not be returned.}
}
\description{
Returns information on the mid object including a summary for encoders, uninterpreted rates, or other special components.
}
\examples{
data(trees, package = "datasets")
mid <- interpret(Volume ~ ., trees, k = 10)
mid.extract(mid, encoders)
mid.extract(mid, intercept)
mid.extract(mid, uninterpreted.rate)
mid.extract(mid, Girth)
}
