% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid_importance.R
\name{mid.importance}
\alias{mid.importance}
\title{Calculate and Visualize MID-based Importance and Breakdown}
\usage{
mid.importance(object, data = NULL, weights = NULL, sort = TRUE, measure = 1L)
}
\arguments{
\item{object}{a mid object.}

\item{data}{data to be used to calculate the importance. If NULL, the fitted matrix is extracted from the mid object. If the number of row equals to one, the mid breakdown will be calculated.}

\item{weights}{a numeric vector of weights.}

\item{sort}{logical. If TRUE, the data.frame will be sorted by magnitude of importance}

\item{measure}{an integer specifying the type of function to evaluate the importance of each effect. Possible values are "1" for the mean absolute effect, "2" for the root mean square effect, and "3" for the median absolute effect.}
}
\value{
\code{mid.importance} returns a data frame for the term importance, measured as the average magnitude of the effect.
}
\description{
Creates a data frame showing the importance of each functional decomposition term.
\code{mid.importance()} returns a object of class "mid.importance", for which methods for \code{ggplot2::autoplot()} and \code{graphics::barplot()} are defined.
}
\examples{
data(airquality, package = "datasets")
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
imp <- mid.importance(mid)
imp
}
