% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midr-package.R
\docType{package}
\name{midr-package}
\alias{midr}
\alias{midr-package}
\title{midr: Create an Interpretable Surrogate of Black-Box ML Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Construct a globally interpretable surrogate of a black-box ML model by decomposing the prediction function of the target model into a set of component functions including the intercept, the main effects and the second-order interactions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midr}
  \item Report bugs at \url{https://github.com/ryo-asashi/midr/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asasihba \email{ryoichi.asashiba@gmail.com}

Authors:
\itemize{
  \item Hirokazu Iwasawa
}

Other contributors:
\itemize{
  \item Reiji Kozuma [contributor]
}

}
\keyword{internal}
