% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midr-package.R
\docType{package}
\name{midr-package}
\alias{midr}
\alias{midr-package}
\title{midr: Learning from Black-Box Models by Maximum Interpretation Decomposition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

midr provides a model-agnostic method to interpret and explain black-box predictive models by creating a globally interpretable surrogate based on Maximum Interpretation Decomposition (MID), a functional decomposition technique that minimizes the squared error between the target model predictions and the predictions of the surrogate model. For the theoretical foundations of MID, please refer to Iwasawa & Matsumori (2025) [Forthcoming]. For technical details of the midr package, please see Asashiba et al. (2025) [Forthcoming].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midr}
  \item \url{https://ryo-asashi.github.io/midr/}
  \item Report bugs at \url{https://github.com/ryo-asashi/midr/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asasihba \email{ryoichi.asashiba@gmail.com}

Authors:
\itemize{
  \item Hirokazu Iwasawa
}

Other contributors:
\itemize{
  \item Reiji Kozuma [contributor]
}

}
\keyword{internal}
