% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midr-package.R
\docType{package}
\name{midr-package}
\alias{midr}
\alias{midr-package}
\title{midr: Interpret Black-Box Models by Maximum Interpretation Decomposition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

midr is designed to provide a model-agnostic method for interpreting and explaining black-box machine learning models by creating a globally interpretable surrogate of the target model using a functional decomposition technique called Maximum Interpretation Decomposition (MID). For the theoretical details of MID, see Iwasawa & Matsumori (2025) <d>[Forthcoming], and for the technical details of the package, see Asashiba et al. (2025) [Forthcoming].
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midr}
  \item \url{https://ryo-asashi.github.io/midr/}
  \item Report bugs at \url{https://github.com/ryo-asashi/midr/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asasihba \email{ryoichi.asashiba@gmail.com}

Authors:
\itemize{
  \item Hirokazu Iwasawa
}

Other contributors:
\itemize{
  \item Reiji Kozuma [contributor]
}

}
\keyword{internal}
