% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midr-package.R
\docType{package}
\name{midr-package}
\alias{midr}
\alias{midr-package}
\title{midr: Learning from Black-Box Models by Maximum Interpretation Decomposition}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'midr' is to provide a model-agnostic method for interpreting and explaining black-box predictive models by creating a globally interpretable surrogate model. The package implements 'Maximum Interpretation Decomposition' (MID), a functional decomposition technique that finds an optimal additive approximation of the original model. This approximation is achieved by minimizing the squared error between the predictions of the black-box model and the surrogate model. The theoretical foundations of MID are described in Iwasawa & Matsumori (2025) [Forthcoming], and the package itself is detailed in Asashiba et al. (2025) \doi{10.48550/arXiv.2506.08338}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ryo-asashi/midr}
  \item \url{https://ryo-asashi.github.io/midr/}
  \item Report bugs at \url{https://github.com/ryo-asashi/midr/issues}
}

}
\author{
\strong{Maintainer}: Ryoichi Asasihba \email{ryoichi.asashiba@gmail.com}

Authors:
\itemize{
  \item Hirokazu Iwasawa
}

Other contributors:
\itemize{
  \item Reiji Kozuma [contributor]
}

}
\keyword{internal}
