% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_encoder.R
\name{numeric.encoder}
\alias{numeric.encoder}
\alias{numeric.frame}
\title{Encoder for Quantitative Variables}
\usage{
numeric.encoder(
  x,
  k,
  type = 1L,
  encoding.digits = NULL,
  tag = "x",
  frame = NULL,
  weights = NULL
)

numeric.frame(
  reps = NULL,
  breaks = NULL,
  type = NULL,
  encoding.digits = NULL,
  tag = "x"
)
}
\arguments{
\item{x}{a numeric vector which is to be encoded.}

\item{k}{an integer specifying the coarseness of the encoding. If not positive, all unique values of x are chosen as sample points.}

\item{type}{an integer specifying the shape of the function to be fit. \code{1} is for a piecewise linear function and \code{0} is for a piecewise constant (step) function.}

\item{encoding.digits}{an integer specifying the rounding digits for encoding numeric variables when \code{type} is 1 (piecewise linear functions).}

\item{tag}{name of the corresponding predictor variable.}

\item{frame}{a \code{numeric.frame} object containing the information about the binning of the variable.}

\item{weights}{optional. a numeric vector indicating the weight of each value of 'x'.}

\item{reps}{a numeric vector specifying the representative values of each bin.}

\item{breaks}{a numeric vector to be used as the breaks of the binning.}
}
\description{
Returns a list consisting of the encoding information of a predictor variable as a quantitative variable.
}
