% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid_conditional.R
\name{plot.mid.conditional}
\alias{plot.mid.conditional}
\title{Plot MID Individual Conditional Expectations with graphics Package}
\usage{
\method{plot}{mid.conditional}(
  x,
  centered = FALSE,
  draw.dots = TRUE,
  sample = NULL,
  term = NULL,
  variable.alpha = NULL,
  variable.colour = NULL,
  variable.linetype = NULL,
  variable.linewidth = NULL,
  scale.palette = NULL,
  ...
)
}
\arguments{
\item{x}{a mid.conditional object to visualize.}

\item{centered}{logical.}

\item{draw.dots}{logical. If TRUE, points representing the predictions at the observed values are}

\item{sample}{a vector specifying the set of names of the observations to be plotted.}

\item{term}{an optional character specifying one of the relevant terms. If passed, the individual conditional expectations for the specified term are plotted.}

\item{variable.alpha}{a name of the predictor variable to use to set \code{alpha} for each plot.}

\item{variable.colour}{a name of the predictor variable to use to set \code{color} for each plot.}

\item{variable.linetype}{a name of the predictor variable to use to set \code{linetype} for each plot.}

\item{variable.linewidth}{a name of the predictor variable to use to set \code{linewidth} for each plot.}

\item{scale.palette}{a character vector of color names, specifying the colors to be used.}

\item{...}{optional parameters to be directly passed to \code{ggplot2::geom_line()}.}
}
\description{
Creates a plot representing the MID-based individual conditional expectations
}
