% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid_importance.R
\name{plot.mid.importance}
\alias{plot.mid.importance}
\title{Plot MID Importance with Basic Functions}
\usage{
\method{plot}{mid.importance}(
  x,
  type = c("barplot", "heatmap", "dotchart", "boxplot"),
  max.bars = NA,
  scale.palette = c("#FFFFFF", "gray10"),
  ...
)
}
\arguments{
\item{x}{a "mid.importance" object to be visualized.}

\item{type}{a character string specifying the type of the plot. One of "barplot", "heatmap", "dotchart" or "boxplot".}

\item{max.bars}{an integer specifying the maximum number of bars in the barplot, boxplot and dotchart.}

\item{scale.palette}{a character vector of length two to be used as the color palette for the heatmap.}

\item{...}{optional parameters to be passed to the graphing function.}
}
\value{
\code{plot.mid.importance()} produces a plot and returns \code{NULL}.
}
\description{
For "mid.importance" objects, \code{plot()} visualizes the importance of MID component functions.
}
\details{
The S3 method of \code{plot()} for "mid.importance" objects creates a visualization of the MID importance using \code{graphics::barplot()}, \code{graphics::image()} or \code{graphics::dotchart()}.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])
imp <- mid.importance(mid)
plot(imp)
plot(imp, type = "heatmap")
plot(imp, type = "dotchart")
}
