% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid.R
\name{plot.mid}
\alias{plot.mid}
\title{Plot MID Component Functions}
\usage{
\method{plot}{mid}(
  x,
  term,
  type = c("effect", "data", "compound"),
  theme = NULL,
  intercept = FALSE,
  main.effects = FALSE,
  data = NULL,
  limits = NULL,
  jitter = 0.3,
  cells.count = c(100L, 100L),
  ...
)
}
\arguments{
\item{x}{a "mid" object to be visualized.}

\item{term}{a character string specifying the component function to be plotted.}

\item{type}{the method for plotting. One of "effect", "data" or "compound".}

\item{theme}{a character string or object defining the color theme.}

\item{intercept}{logical. If \code{TRUE}, the intercept is added to the MID values.}

\item{main.effects}{logical. If \code{TRUE}, main effects are included in the interaction plot.}

\item{data}{a data frame to be plotted with the corresponding MID values. If not provided, data is automatically extracted from the "mid" object.}

\item{limits}{a numeric vector of length two specifying the limits of the plotting scale.}

\item{jitter}{a numeric value specifying the amount of jitter for the data points.}

\item{cells.count}{an integer or vector of two integers specifying the resolution of the raster plot for interactions.}

\item{...}{optional parameters to be passed to the graphing function. Possible arguments are "col", "fill", "pch", "cex", "lty", "lwd" and aliases of them.}
}
\value{
\code{plot.mid()} produces a plot as a side-effect and returns \code{NULL} invisibly.
}
\description{
\code{plot.mid()} is an S3 method for "mid" objects that visualizes a single component function using base R's graphing functions.
}
\details{
The \code{type} argument controls the plotting mode.
When \code{type = "effect"} (default), the component function itself is plotted.
The plotting type is automatically selected based on the effect's type:
for a quantitative main effect, a line plot is drawn;
for a qualitative main effect, a bar plot is drawn; and
for an interaction effect, a raster plot is drawn.
If \code{type = "data"}, only the values of the component function at the raw data points are plotted, which requires the \code{data} argument to be supplied.
Finally, \code{type = "compound"} combines both modes, plotting the component function as well as the raw data points.
}
\examples{
data(diamonds, package = "ggplot2")
set.seed(42)
idx <- sample(nrow(diamonds), 1e4)
mid <- interpret(price ~ (carat + cut + color + clarity)^2, diamonds[idx, ])

# Plot a quantitative main effect
plot(mid, "carat")

# Plot a qualitative main effect
plot(mid, "clarity")

# Plot an interaction effect with data points and a raster layer
plot(mid, "carat:clarity", type = "compound", data = diamonds[idx, ])

# Use a different colot theme
plot(mid, "clarity:color", theme = "RdBu")
}
\seealso{
\code{\link{interpret}}, \code{\link{ggmid}}
}
