% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mid.R
\name{plot.mid}
\alias{plot.mid}
\title{Plot MID with Basic Functions}
\usage{
\method{plot}{mid}(
  x,
  term,
  add.intercept = FALSE,
  include.main.effects = FALSE,
  scale.type = "default",
  scale.palette = c("#2f7a9a", "#FFFFFF", "#7e1952"),
  cells.count = c(100L, 100L),
  ...
)
}
\arguments{
\item{x}{a "mid" object to be visualized.}

\item{term}{a character string specifying the component function to be plotted.}

\item{add.intercept}{logical. If \code{TRUE}, the intercept is added to the MID values and the plotting scale is shifted.}

\item{include.main.effects}{logical. If \code{TRUE}, the main effects are included in the interaction plot.}

\item{scale.type}{character string. The color type of the interaction plot. One of "default", "viridis", "gradient".}

\item{scale.palette}{a character vector of color names specifying the colors to be used in the interaction plot.}

\item{cells.count}{an integer or integer-valued vector of length two specifying the number of cells for the raster type interaction plot.}

\item{...}{optional parameters to be passed to the graphing function.}
}
\value{
\code{plot.mid()} produces a line plot or bar plot for a main effect and a filled contour plot for an interaction and returns \code{NULL}.
}
\description{
For "mid" objects, \code{plot()} visualizes a MID component function.
}
\details{
The S3 method of \code{plot()} for "mid" objects creates a visualization of a MID component function using \code{base::plot()} for a main effect of a quantitative variable, \code{graphics::barplot()} for a main effect of a qualitative variable, and \code{graphics::filled.contour()} for an interaction.
}
\examples{
data(airquality, package = "datasets")
airquality$Month <- factor(airquality$Month)
mid <- interpret(Ozone ~ .^2, airquality, lambda = 1)
plot(mid, "Temp")
plot(mid, "Month")
plot(mid, "Wind:Temp")
plot(mid, "Solar.R:Month", scale.type = "viridis",
     add.intercept = TRUE, include.main.effects = TRUE)
}
