% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mid.R
\name{predict.mid}
\alias{predict.mid}
\alias{mid.f}
\title{Predict Method for MID Model}
\usage{
\method{predict}{mid}(
  object,
  newdata = NULL,
  na.action = getOption("na.action"),
  type = c("response", "link", "terms"),
  terms = object$terms,
  ...
)

mid.f(object, term, x, y = NULL)
}
\arguments{
\item{object}{a 'mid' object to be used to make predictions.}

\item{newdata}{a data frame of the new observations.}

\item{na.action}{a function which indicates what should happen when the data contain \code{NA}s.}

\item{type}{the type of prediction required. The default is on the scale of the response varialbe. The alternative "link" is on the scale of the linear predictors. The "terms" option returns a matrix giving the fitted values of each term in the model formula on the linear predictor scale.}

\item{terms}{a character vector specifying the subset of component functions to be used for prediction.}

\item{...}{not used.}

\item{term}{a character specifying the component function of a fitted MID model.}

\item{x}{a vector to be used as the input to the first argument of the component function. If a matrix or data frame is passed, inputs are extracted from the matrix or data frame.}

\item{y}{a vector to be used as the input to the second argument of the component function.}
}
\value{
\code{predict.mid()} returns a numeric vector of MID model predictions.
}
\description{
The method of \code{predict()} for 'mid' objects obtains predictions from a fitted MID model.
}
\details{
The S3 method of \code{predict()} for MID models returns the model predictions.
\code{mid.f()} works as a component function of a MID model.
}
\examples{
data(cars, package = "datasets")
mid <- interpret(dist ~ speed, cars, lambda = 1)
predict(mid, newdata = data.frame(speed = 5:25))
mid.f(mid, "speed", 5:25)
mid.f(mid, "speed", 5:25) + mid$intercept
}
