% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapviz_mid.R
\name{shapviz.mid}
\alias{shapviz.mid}
\title{Calculate SHAP of MID Predictions}
\usage{
\method{shapviz}{mid}(object, data = NULL)
}
\arguments{
\item{object}{a "mid" object.}

\item{data}{a data frame containing observations for which SHAP values are calculated. If not passed, data is extracted from \code{parent.env()} based on the function call of the "mid" object.}
}
\value{
\code{shapviz.mid()} returns an object of class "shapviz".
}
\description{
\code{shapviz.mid()} is a S3 method of \code{shapviz::shapviz()} for the fitted MID models.
}
\details{
The S3 method of \code{shapviz()} for the "mid" objects returns an object of class "shapviz" to be used to create SHAP plots with the functions of the shapviz package such as \code{sv_waterfall()} and \code{sv_importance()}.
}
