% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_mid.R
\name{summary.mid}
\alias{summary.mid}
\title{Summary Method for MID Objects}
\usage{
\method{summary}{mid}(object, digits = max(3L, getOption("digits") - 2L), top.n = 10L, ...)
}
\arguments{
\item{object}{a mid object to be printed.}

\item{digits}{number of significant digits.}

\item{top.n}{an integer specifying the maximum number of terms to be displayed with importance values.}

\item{...}{not used.}
}
\description{
Prints the summary of the fitted mid object.
}
\examples{
data(cars, package = "datasets")
print(interpret(dist ~ speed, cars))
}
