% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{theme_midr}
\alias{theme_midr}
\title{Themes for ggplot Objects}
\usage{
theme_midr(
  grid_type = c("none", "x", "y", "xy"),
  base_size = 11,
  base_family = "serif",
  base_line_size = base_size/22,
  base_rect_size = base_size/22
)
}
\arguments{
\item{grid_type}{one of "none", "x", "y" or "both".}

\item{base_size}{a positive value for the base font size, given in pts.}

\item{base_family}{a character specifying the base font family.}

\item{base_line_size}{a positive value for the base size for line elements.}

\item{base_rect_size}{a positive value for the base size for rect elements.}
}
\value{
\code{theme_midr()} provides a ggplot2 theme customized for the midr package.
}
\description{
Returns a complete theme for \code{ggplot} objects.
}
\examples{
ggplot2::theme_set(theme_midr())
}
