% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted}
\alias{weighted}
\alias{augmented}
\alias{shuffled}
\alias{latticized}
\alias{weights.weighted}
\title{Weighted Data Frames}
\usage{
weighted(data, weights = NULL, ...)

augmented(data, weights = NULL, size = nrow(data), ratio = 0.01)

shuffled(data, weights = NULL, size = nrow(data))

latticized(
  data,
  weights = NULL,
  k = 10L,
  type = 1L,
  use.catchall = TRUE,
  catchall = "(others)",
  frames = list(),
  keep.mean = TRUE
)

\method{weights}{weighted}(object, ...)
}
\arguments{
\item{data}{a data frame for which weights are to be added.}

\item{weights}{a numeric vector, specifying the weights of each record.}

\item{...}{not used.}

\item{size}{an integer, specifying the number of new rows to be added to the original data frame.}

\item{ratio}{a numeric value. Weights for the new rows are calculated as \code{nrow(data) * ratio / size}.}

\item{k}{an integer or a numeric vector of length two for main effects and interactions, specifying the maximum number of sample points for each numeric predictor variable. If an integer is passed, k is used for main effect terms and the square root of k is used for interaction terms. If not positive, all unique values are used as sample points.}

\item{type}{an integer or a vector of length two, specifying the type of piecewise functions to be fit on numeric variables. '0' is for step functions on discretised intervals, and '1' is for piecewise linear functions connecting at representative values.}

\item{use.catchall}{logical. If TRUE, less frequent levels of factor variables are dropped and replaced with the catchall level.}

\item{catchall}{a character used as the name of "catchall" level for unused levels of each factor variable. Used only when \code{factor.option} is not 0.}

\item{frames}{a named list of encoding frames, which specifies bins for quantitative features or levels for qualitative features.}

\item{keep.mean}{logical. If TRUE, the mean is used}

\item{object}{a weighted data frame.}
}
\description{
Returns a data frame of class \code{weighted} and \code{data.frame}, which contains the additional attribute "weights".
Weights can be extracted using functions like \code{attr()} or \code{stats::weights()}.
}
\examples{
x1 <- runif(1000L, -1, 1)
x2 <- x1 + runif(1000L, -1, 1)
X <- weighted(cbind(x1, x2), (abs(x1) + abs(x2)) / 2)
Y1 <- weighted(X)
ggplot2::ggplot(Y1, ggplot2::aes(x1, x2, alpha = weights(Y1))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "weighted") +
  ggplot2::theme_bw()
Y2 <- shuffled(X)
ggplot2::ggplot(Y2, ggplot2::aes(x1, x2, alpha = weights(Y2))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "shuffled") +
  ggplot2::theme_bw()
Y3 <- augmented(X)
ggplot2::ggplot(Y3, ggplot2::aes(x1, x2, alpha = weights(Y3))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "augmented") +
  ggplot2::theme_bw()
Y4 <- latticized(X)
ggplot2::ggplot(Y4, ggplot2::aes(x1, x2, size = weights(Y4))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "latticized") +
  ggplot2::theme_bw()
}
