% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.R
\name{weighted}
\alias{weighted}
\alias{augmented}
\alias{shuffled}
\alias{latticized}
\alias{weights.weighted}
\title{Weighted Data Frames}
\usage{
weighted(data, weights = NULL, ...)

augmented(data, weights = NULL, size = nrow(data), r = 0.01)

shuffled(data, weights = NULL, size = nrow(data))

latticized(
  data,
  weights = NULL,
  k = 10L,
  type = 0L,
  use.catchall = TRUE,
  catchall = "(others)",
  frames = list(),
  keep.mean = TRUE
)

\method{weights}{weighted}(object, ...)
}
\arguments{
\item{data}{a data frame.}

\item{weights}{a numeric vector of sample weights for each observation in \code{data}.}

\item{...}{not used.}

\item{size}{integer. The number of random observations whose values are sampled from the marginal distribution of each variable.}

\item{r}{a numeric value specifying the ratio of the total weights for the random observations to the sum of sample weights. The weight for the random observations is calculated as \code{sum(attr(data, "weights")) * r / size}.}

\item{k}{integer. The maximum number of sample points for each variable. If not positive, all unique values are used as sample points.}

\item{type}{integer. The type of encoding of quantitative variables to be passed to \code{numeric.encoder()}.}

\item{use.catchall}{logical. If \code{TRUE}, less frequent levels of factor variables are dropped and replaced by the catchall level.}

\item{catchall}{a character string to be used as the catchall level.}

\item{frames}{a named list of encoding frames ("numeric.frame" or "factor.frame" objects).}

\item{keep.mean}{logical. If \code{TRUE}, the representative values of each group is the average of the corresponding group.}

\item{object}{a data frame with the attribute "weights".}
}
\value{
\code{weighted()} returns a data frame with the attribute "weights".
\code{augmented()} returns a weighted data frame of the original data and the shuffled data with relatively small weights.
\code{shuffled()} returns a weighted data frame of the shuffled data.
\code{latticized()} returns a weighted data frame of latticized data, whose values are grouped and replaced by the representative value of the corresponding group.
}
\description{
\code{weighted()} returns a data frame with sample weights.
}
\details{
\code{weighted()} returns a data frame with the "weights" attribute that can be extracted using \code{stats::weights()}.
\code{augmented()}, \code{shuffled()} and \code{latticized()} return a weighted data frame with some data modifications.
These functions are designed for use with \code{interpret()}.
As the modified data frames do not preserve the original correlation structure of the variables, the response variable (y) should always be replaced by the model predictions (yhat).
}
\examples{
x1 <- runif(1000L, -1, 1)
x2 <- x1 + runif(1000L, -1, 1)
X <- weighted(cbind(x1, x2), (abs(x1) + abs(x2)) / 2)
Y1 <- weighted(X)
ggplot2::ggplot(Y1, ggplot2::aes(x1, x2, alpha = weights(Y1))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "weighted")
Y2 <- shuffled(X)
ggplot2::ggplot(Y2, ggplot2::aes(x1, x2, alpha = weights(Y2))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "shuffled")
Y3 <- augmented(X)
ggplot2::ggplot(Y3, ggplot2::aes(x1, x2, alpha = weights(Y3))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "augmented")
Y4 <- latticized(X)
ggplot2::ggplot(Y4, ggplot2::aes(x1, x2, size = weights(Y4))) +
  ggplot2::geom_point() +
  ggplot2::labs(title = "latticized")
}
